Split Bundle and RelationDefinition to separate modules. Complex merge_bundles() implementation.
