+++ .pre-commit-config.yaml
@@ -20,7 +20,7 @@ repos:
   hooks:
     - id: ty-check
       name: ty-check
-      language: python
+      language: system
       entry: uv run --no-sync ty check
       types: [python]
       pass_filenames: true
