+++ lynxkite-graph-analytics/src/lynxkite_graph_analytics/core.py
@@ -10,15 +10,18 @@ import dataclasses
 import functools
 import networkx as nx
 import pandas as pd
-import polars as pl
 import traceback
 import typing
 import urllib.parse
+
+from .relation_definition import RelationDefinition  # noqa: F401
+from .bundle import Bundle
 
 if typing.TYPE_CHECKING:
     import fastapi
 
 ENV = "LynxKite Graph Analytics"
+
 
 # Annotated types with format "dropdown" let you specify the available options
 # as a query on the input_metadata. These query expressions are JMESPath expressions.
@@ -78,176 +81,6 @@ DataFrameColumn = typing.Annotated[
 """A type annotation to be used for parameters of an operation. DataFrameColumn is
 rendered as a dropdown in the frontend, listing the columns of the "df" DataFrame.
 The column name is passed to the operation as a string."""
-
-
-@dataclasses.dataclass
-class RelationDefinition:
-    """
-    Defines a set of edges.
-
-    Attributes:
-        df: The name of the DataFrame that contains the edges.
-        source_column: The column in the edge DataFrame that contains the source node ID.
-        target_column: The column in the edge DataFrame that contains the target node ID.
-        source_table: The name of the DataFrame that contains the source nodes.
-        target_table: The name of the DataFrame that contains the target nodes.
-        source_key: The column in the source table that contains the node ID.
-        target_key: The column in the target table that contains the node ID.
-        name: Descriptive name for the relation.
-    """
-
-    df: str
-    source_column: str
-    target_column: str
-    source_table: str
-    target_table: str
-    source_key: str
-    target_key: str
-    name: str
-
-
-@dataclasses.dataclass
-class Bundle:
-    """A collection of DataFrames and other data.
-
-    Can efficiently represent a knowledge graph (homogeneous or heterogeneous) or tabular data.
-
-    By convention, if it contains a single DataFrame, it is called `df`.
-    If it contains a homogeneous graph, it is represented as two DataFrames called `nodes` and
-    `edges`.
-
-    Attributes:
-        dfs: Named DataFrames.
-        relations: Metadata that describes the roles of each DataFrame.
-            Can be empty, if the bundle is just one or more DataFrames.
-        other: Other data, such as a trained model.
-    """
-
-    dfs: dict[str, pd.DataFrame] = dataclasses.field(default_factory=dict)
-    relations: list[RelationDefinition] = dataclasses.field(default_factory=list)
-    other: dict[str, typing.Any] = dataclasses.field(default_factory=dict)
-
-    @classmethod
-    def from_nx(cls, graph: nx.Graph):
-        edges = nx.to_pandas_edgelist(graph)
-        d = dict(graph.nodes(data=True))
-        nodes = pd.DataFrame(d.values(), index=list(d.keys()))
-        nodes["id"] = nodes.index
-        if "index" in nodes.columns:
-            nodes.drop(columns=["index"], inplace=True)
-        return cls(
-            dfs={"edges": edges, "nodes": nodes},
-            relations=[
-                RelationDefinition(
-                    name="edges",
-                    df="edges",
-                    source_column="source",
-                    target_column="target",
-                    source_table="nodes",
-                    target_table="nodes",
-                    source_key="id",
-                    target_key="id",
-                )
-            ],
-        )
-
-    @classmethod
-    def from_df(cls, df: pd.DataFrame):
-        return cls(dfs={"df": df})
-
-    def to_nx(self):
-        # TODO: Use relations.
-        graph = nx.DiGraph()
-        if "nodes" in self.dfs:
-            df = self.dfs["nodes"]
-            if df.index.name != "id":
-                df = df.set_index("id")
-            graph.add_nodes_from(df.to_dict("index").items())
-        if "edges" in self.dfs:
-            edges = self.dfs["edges"]
-            graph.add_edges_from(
-                [
-                    (
-                        e["source"],
-                        e["target"],
-                        {k: e[k] for k in edges.columns if k not in ["source", "target"]},
-                    )
-                    for e in edges.to_records()
-                ]
-            )
-        return graph
-
-    def copy(self):
-        """
-        Returns a shallow copy of the bundle. The Bundle and its containers are new, but
-        the DataFrames and RelationDefinitions are shared. (The contents of `other` are also shared.)
-        """
-        return Bundle(
-            dfs=dict(self.dfs),
-            relations=list(self.relations),
-            other=dict(self.other),
-        )
-
-    def metadata(self):
-        """JSON-serializable information about the bundle, metadata only."""
-        return {
-            "dataframes": {
-                name: {
-                    "key": name,
-                    "columns": sorted(str(c) for c in df.columns),
-                }
-                for name, df in self.dfs.items()
-            },
-            "relations": [dataclasses.asdict(relation) for relation in self.relations],
-            "other": {
-                k: {"key": k, **getattr(v, "metadata", lambda: {})()} for k, v in self.other.items()
-            },
-        }
-
-    def to_table_view(self, limit: int = 100):
-        """Converts the bundle to a format suitable for display as tables in the frontend."""
-        return BundleTableView.from_bundle(self, limit=limit)
-
-    def merge(self, other: "Bundle"):
-        """Merges another bundle into this one. Modifies this bundle in-place."""
-        self.dfs.update(other.dfs)
-        self.relations.extend(other.relations)
-        self.other.update(other.other)
-
-
-@dataclasses.dataclass
-class SingleTableView:
-    """A JSON-serializable view of a table in the bundle, for use in the frontend.
-
-    Attributes:
-        columns: The columns to display.
-        data: A list of rows, where each row is a list of values.
-    """
-
-    columns: list[str]
-    data: list[list[typing.Any]]
-
-    @staticmethod
-    def from_df(df: pd.DataFrame, limit: int = 100):
-        columns = [str(c) for c in df.columns][:limit]
-        df = df[columns]
-        data = df_for_frontend(df, limit).values.tolist()
-        return SingleTableView(columns=columns, data=data)
-
-
-@dataclasses.dataclass
-class BundleTableView:
-    """A JSON-serializable tabular view of a bundle, for use in the frontend."""
-
-    dataframes: dict[str, SingleTableView]
-    relations: list[RelationDefinition]
-    other: dict[str, typing.Any]
-
-    @staticmethod
-    def from_bundle(bundle: Bundle, limit: int = 100):
-        dataframes = {name: SingleTableView.from_df(df, limit) for name, df in bundle.dfs.items()}
-        other = {k: str(v)[:limit] for k, v in bundle.other.items()}
-        return BundleTableView(dataframes=dataframes, relations=bundle.relations, other=other)
 
 
 def nx_node_attribute_func(name):
@@ -484,20 +317,6 @@ def _get_metadata(x) -> dict:
     if hasattr(x, "metadata"):
         return x.metadata()
     return {}
-
-
-def df_for_frontend(df: pd.DataFrame, limit: int) -> pd.DataFrame:
-    """Returns a DataFrame with values that are safe to send to the frontend."""
-    df = df[:limit]
-    if isinstance(df, pl.LazyFrame):
-        df = df.collect()
-    if isinstance(df, pl.DataFrame):
-        df = df.to_pandas()
-    # Convert non-numeric columns to strings.
-    for c in df.columns:
-        if not pd.api.types.is_numeric_dtype(df[c]):
-            df[c] = df[c].astype(str)
-    return df
 
 
 async def get_node_service(request: "fastapi.Request") -> Service:
